Refactor make_credential
