+++ src/fidokey/make_credential/make_credential_command.rs
@@ -1,5 +1,5 @@
 use crate::ctapdef;
-use crate::make_credential_params::{CredentialSupportedKeyType, Extension};
+use super::make_credential_params::{CredentialSupportedKeyType, Extension};
 use crate::util;
 use serde_cbor::to_vec;
 use serde_cbor::Value;
